feat: new logic supports full content transformation
