+++ transform_test.go
@@ -96,48 +96,6 @@ func TestTransformPage(t *testing.T) {
 		require.Equal(t, filepath.Join("content", "posts", "name-with-space.md"), result.filename)
 	})
 
-	t.Run("removes dashes with no text after them", func(t *testing.T) {
-		result := transformText("-\n\t- \n\t\t-")
-		require.Equal(t, "\n\n", result)
-	})
-
-	t.Run("removes dashes from the text", func(t *testing.T) {
-		result := transformText("-\n- hello")
-		require.Equal(t, "\n\nhello", result)
-	})
-
-	t.Run("turns second level bullet points into first level", func(t *testing.T) {
-		result := transformText("\t- hello\n\t- world")
-		require.Equal(t, "\n- hello\n\n- world", result) // TODO: maybe remove the duplicated new line
-	})
-
-	t.Run("removes one tab from multi-level bullet points", func(t *testing.T) {
-		result := transformText("\t\t- hello\n\t\t\t- world")
-		require.Equal(t, "\t- hello\n\t\t- world", result)
-	})
-	t.Run("removes tabs from all subsequent lines of a bullet point", func(t *testing.T) {
-		result := transformText(`
-- ~~~ts
-  const hello = "world";
-  ~~~
-- single line
-- multiple
-  lines
-  in
-  one`)
-		require.Equal(t, `
-
-~~~ts
-const hello = "world";
-~~~
-
-single line
-
-multiple
-lines
-in
-one`, result)
-	})
 }
 
 func TestTransformImages(t *testing.T) {
@@ -148,7 +106,7 @@ func TestTransformImages(t *testing.T) {
 		}
 		result := transformPage(testPage, "/images")
 		require.Equal(t, []string{"../assets/image.png"}, result.assets)
-		require.Equal(t, "\n![hello world](/images/image.png)", result.text)
+		require.Equal(t, "- ![hello world](/images/image.png)", result.text)
 	})
 
 	t.Run("ignores absolute images", func(t *testing.T) {
@@ -158,7 +116,7 @@ func TestTransformImages(t *testing.T) {
 		}
 		result := transformPage(testPage, "/images")
 		require.Equal(t, 0, len(result.assets))
-		require.Equal(t, "\n![hello world](http://example.com/assets/image.png)", result.text)
+		require.Equal(t, "- ![hello world](http://example.com/assets/image.png)", result.text)
 	})
 
 	t.Run("extracts relative images from image attribute", func(t *testing.T) {
